feat(perf): code-split MDX client components into individual files
