+++ src/components/mdx/client/Tabs.tsx
@@ -1,0 +1,34 @@
+"use client";
+
+import React from "react";
+
+interface TabsProps {
+  tabs: { id: string; label: string; content: React.ReactNode }[];
+}
+
+export function Tabs({ tabs }: TabsProps) {
+  const [activeTab, setActiveTab] = React.useState(tabs[0]?.id || "");
+
+  return (
+    <div className="my-6 not-prose">
+      <div className="flex border-b border-border overflow-x-auto">
+        {tabs.map((tab) => (
+          <button
+            key={tab.id}
+            onClick={() => setActiveTab(tab.id)}
+            className={`px-4 py-2 text-sm font-medium whitespace-nowrap transition-colors ${
+              activeTab === tab.id
+                ? "text-primary border-b-2 border-primary -mb-px"
+                : "text-muted-foreground hover:text-foreground"
+            }`}
+          >
+            {tab.label}
+          </button>
+        ))}
+      </div>
+      <div className="py-4">
+        {tabs.find((tab) => tab.id === activeTab)?.content}
+      </div>
+    </div>
+  );
+}
