+++ src/components/mdx/index.tsx
@@ -5,11 +5,12 @@
  * maintaining backward compatibility while enabling proper RSC architecture.
  *
  * Server components are in server-components.tsx (no "use client")
- * Client components are in client-components.tsx (with "use client")
+ * Client components are split into individual files in client/ directory,
+ * each with its own "use client" boundary for per-component code-splitting.
  */
 
 import { mdxServerComponents } from "./server-components";
-import { mdxClientComponents } from "./client-components";
+import { mdxClientComponents } from "./client";
 
 /**
  * Combined MDX components object
@@ -22,4 +23,4 @@ export const mdxComponents = {
 
 // Re-export individual components for direct imports (backward compatibility)
 export * from "./server-components";
-export * from "./client-components";
+export * from "./client";
